+++ src/mailman/model/tests/test_bans.py
@@ -46,3 +46,17 @@ class TestMailingListBans(unittest.TestCase):
         getUtility(IListManager).delete(self._mlist)
         self.assertEqual([ban.email for ban in global_ban_manager],
                          ['bart@example.com'])
+
+    def test_bans_sequence(self):
+        # Bans returns a pageable sequence.
+        self._manager.ban('bee@example.com')
+        self._manager.ban('ant@example.com')
+        # The results can be len()'d.
+        self.assertEqual(len(self._manager.bans), 2)
+        # The results can be iterated.
+        self.assertEqual(['ant@example.com', 'bee@example.com'],
+                         sorted(ban.email for ban in self._manager.bans))
+        # The results can be indexed.
+        self.assertEqual(
+            sorted([self._manager.bans[0].email, self._manager.bans[1].email]),
+            ['ant@example.com', 'bee@example.com'])